fix: extended turns weren't preserved nor applied
This is caused because if a turn wasn't extended, nor it was part of the previous big junction we always returned the default turn. This fix introduces a new metadata to keep track of extended turns, and if one is met which isn't extended by any other means - we'll return the extended turn rather than the default one.